social-media/backend/requirements.txt
django
django-ninja
django-cors-headers
djangorestframework-simplejwt
uvicorn
django-ninja-extra
django-ninja-jwt